requirements.txt
streamlit
flask
transformers
torch
numpy
pandas
scipy
matplotlib
pinecone-client
langchain
openai-whisper
youtube-transcript-api
pypdf2
tesseract
pytesseract
opencv-python
Pillow
langchain_community
google-generativeai
python-dotenv
pathlib
sentencepiece


